Add Node.js domains support
A connection pool acts as a resource manager and for users such
operations as `acquire` and `release` should behave as `new` and
`delete`. That means, if domains are involved, acquired resources should
be reassigned to an active domain which is a `process.domain` at the
time of `acquire` request, not response. The opposite operation should
also happen when resource is being released, it should be unassigned
from its domain if such exists.